MAGECLOUD-2047: Add Variable to Configure Database Connection Settings
