+++ src/Config/Stage/DeployInterface.php
@@ -16,6 +16,7 @@ interface DeployInterface extends StageConfigInterface
     const VAR_SEARCH_CONFIGURATION = 'SEARCH_CONFIGURATION';
     const VAR_CACHE_CONFIGURATION = 'CACHE_CONFIGURATION';
     const VAR_SESSION_CONFIGURATION = 'SESSION_CONFIGURATION';
+    const VAR_DATABASE_CONFIGURATION = 'DATABASE_CONFIGURATION';
     const VAR_CRON_CONSUMERS_RUNNER = 'CRON_CONSUMERS_RUNNER';
     const VAR_CLEAN_STATIC_FILES = 'CLEAN_STATIC_FILES';
     const VAR_STATIC_CONTENT_SYMLINK = 'STATIC_CONTENT_SYMLINK';
